refactor: Phase 3.1c — extract VisualControlsMixin from ElectrodeAdvisorWidget
Move visual controls panel creation (397 lines) to VisualControlsMixin
in main_window_visual_controls.py. Also decomposed the monolithic
_create_visual_controls_panel into 10 focused sub-methods.

Combined with DrawingMixin and InputPanelMixin, the monolith drops
from 4,387 to 3,150 lines (-28.2%).

Extracted mixins:
- DrawingMixin (510 lines) - 3D rendering/path drawing
- InputPanelMixin (402 lines) - input parameter controls
- VisualControlsMixin (485 lines) - visual controls panel
